requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==45.0.4
Django==5.2.2
djangorestframework==3.16.0
pycparser==2.22
sqlparse==0.5.3
